Split Name::Notify tests out of name_test.rb (#4708)
Moves the 6 tests exercising Name::Notify (email notification
heuristics, Interest state=false opt-out, skip_notify,
classification-only-save guard, notify_webmaster) into
test/models/name/notify_test.rb, matching app/models/name/notify.rb.
Needs `include ActiveJob::TestHelper` (for assert_enqueued_jobs/
assert_no_enqueued_jobs) same as the parent NameTest class had.

Also moves test_has_eol_data and test_has_eol_data_true_via_vetted_description
into Name::TaxonomyTest - Name#has_eol_data? is defined in
app/models/name/taxonomy.rb, not notify.rb, despite living right next
to the notification tests in the old file's informal grouping. Missed
in the earlier Taxonomy split since these two tests lived outside
that split's section boundaries.

require_relative'd from name_test.rb so running that file alone still
runs the full original suite.
